Add unary expressions, move around existing ones.
