+++ src/com/expressions/unary/Not.java
@@ -1,0 +1,14 @@
+package com.expressions.unary;
+
+import com.expressions.Expression;
+
+public class Not extends UnaryExpression {
+    public Not(Expression expression) {
+        super(expression);
+    }
+
+    @Override
+    public String getOperator() {
+        return "!";
+    }
+}
